refactor: restructure app files into app/ directory
Move app controller, module, and service into dedicated app/ subdirectory
for better organization and separation of concerns.
