+++ apps/auth-service/src/app/app.module.ts
