+++ apps/auth-service/src/app/app.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@nestjs/common';
+import type { LoginDto } from '@repo/types';
+
+@Injectable()
+export class AppService {
+  login(Credential: LoginDto): string {
+    return `Hello ${Credential.email}!`;
+  }
+}
